requirements.txt
# Agent Skills Server - Dependencies
# ===================================

# Web Framework
starlette>=0.32.0
uvicorn[standard]>=0.24.0

# HTTP Client (built-in urllib, but certifi for SSL)
certifi>=2023.7.22

# Database
# SQLite is built-in, no extra dependency needed

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0

# Optional: Better logging
python-json-logger>=2.0.7

# 文档导出 (PDF / Word / PPT，纯 Python，无系统 cairo 依赖)
markdown>=3.5.0
reportlab>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
